+++ services/course-material/src/components/ContentRenderer/core/embeds/variants/YoutubeEmbedBlock.tsx
@@ -1,0 +1,50 @@
+import { css } from "@emotion/css"
+import { useTranslation } from "react-i18next"
+
+import { EmbedAttributes } from "../../../../../../types/GutenbergBlockAttributes"
+import BreakFromCentered from "../../../../../shared-module/components/Centering/BreakFromCentered"
+import aspectRatioFromClassName from "../../../../../utils/aspectRatioFromClassName"
+
+export const YoutubeEmbedBlock: React.FC<EmbedAttributes> = (props) => {
+  const { t } = useTranslation()
+  const { url } = props
+  let video = url?.split("v=")[1]
+  if (url) {
+    try {
+      const parsedUrl = new URL(url)
+      const vValue = parsedUrl.searchParams.get("v")
+      if (vValue) {
+        video = vValue
+      }
+    } catch (e) {
+      // eslint-disable-next-line i18next/no-literal-string
+      console.error(`Could not parse Youtube url: `, e)
+    }
+  }
+
+  return (
+    <BreakFromCentered sidebar={false}>
+      <div
+        className={css`
+          width: 100%;
+          max-width: 1000px;
+          margin: 0 auto;
+        `}
+      >
+        <iframe
+          className={css`
+            display: block;
+            width: 100%;
+            aspect-ratio: ${aspectRatioFromClassName(props.className)};
+            margin: 4rem 0;
+          `}
+          src={`https://www.youtube-nocookie.com/embed/${video}`}
+          title={t("title-youtube-video-player")}
+          frameBorder="0"
+          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+          allowFullScreen
+        ></iframe>
+      </div>
+    </BreakFromCentered>
+  )
+}
